requirements.txt
pytest
selene>=2.0.0rc9
selenium>=4.0.0,<4.15.0
allure-pytest
webdriver-manager